requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
sentence-transformers
geopy
sqlalchemy
python-dotenv
